CLI fully functional; Enabled sophisticated config file handling
Moved more configs to the config file (e.g. mc-nr-issues, mc-nr-days,
auth)
Enabled default configs within config file. When a configuration
parameter is given outside of a specific project config, it will be
used as default value when a project config doesn't mention it.
So, in a config.edn file one only has to mention base-url once, to use
it in all given project configs. For example:
{:base-url "https://example.com"
"project key" {:issue-query "some jql query here}}

There is an issue with keywords as project keys, because the cli
params are given as strings. Therefore it would mean handling keywords
manually (find out if string begins with : and if so extract the text
behind it to create a keyword). -> TODO